+++ src/org/soulspace/overarch/adapter/ui/cli.clj
@@ -53,6 +53,7 @@
    ["-s" "--select-elements CRITERIA" "Select and print model elements by criteria" :parse-fn edn/read-string]
    ["-S" "--select-references CRITERIA" "Select model elements by criteria and print as references" :parse-fn edn/read-string]
    [nil  "--select-views CRITERIA" "Select and print views by criteria" :parse-fn edn/read-string]
+   [nil  "--select-view-references CRITERIA" "Select views by criteria and print as references" :parse-fn edn/read-string]
    ["-T" "--template-dir DIRNAME" "Template directory" :default "templates"]
    ["-g" "--generation-config FILE" "Generation configuration"]
    ["-G" "--generation-dir DIRNAME" "Generation artifact directory" :default "generated"]
@@ -183,7 +184,7 @@
   (when-let [criteria (spec/check-selection-criteria (:select-references options))]
     (into []
           (comp (model/filter-xf @repo/state criteria)
-                   (map el/element->ref))
+                (map el/element->ref))
           (repo/model-elements))))
 
 (defn select-views
@@ -191,6 +192,14 @@
   [options]
   (when-let [criteria (spec/check-selection-criteria (:select-views options))]
     (into #{} (model/filter-xf @repo/state criteria)
+          (repo/views))))
+
+(defn select-view-references
+  "Returns references to the views selected by criteria specified in the `options`."
+  [options]
+  (when-let [criteria (spec/check-selection-criteria (:select-view-references options))]
+    (into [] (comp (model/filter-xf @repo/state criteria)
+                   (map el/element->ref))
           (repo/views))))
 
 (defn dispatch
@@ -211,6 +220,9 @@
   (when (:select-views options)
     (println "Selected Views for" (:select-views options))
     (pp/pprint (select-views options)))
+  (when (:select-view-references options)
+    (println "Selected Views for" (:select-view-references options))
+    (pp/pprint (select-view-references options)))
   (when (:plantuml-list-sprites options)
     (print-sprite-mappings))
   (when (:render-format options)