+++ Changelog.md
@@ -5,6 +5,7 @@ Version 0.26.0
 --------------
 * *breaking* renamed `:class-view` to `:code-view` to make it more general
 * renamed class model to code model
+* added a `--select-view-references CRITERIA` command line argument
 * added `:per-namespace` artifact generation
 * changed default direction for `:include` relations in plantuml rendering of use case views
 * changed keys in model-info result map to more meaningful names